updated on mixed precision

+++ flash_attention_tensor_cores/testing.cu
@@ -134,7 +134,7 @@ int main(){
     shared_mem_needed += D * tile * sizeof(float); // output
     shared_mem_needed += D * tile * sizeof(__half); // keys + values idepdendently
     shared_mem_needed += tile * tile * sizeof(float); // tile
-    shared_mem_needed += 64 * sizeof(__half); // running max
+    shared_mem_needed += 64 * sizeof(float); // running max
     shared_mem_needed += 64 * sizeof(float); // running sum
